Support release-based versioning and filtering
Versions can be specified by exact version or build. They can also be specified as a major version and release type (prototype, development, testing, stable). The release type defaults to stable, but if you want to install development releases, you can specify one of the development release types. Anything matching that type or newer will be used as the version based on the specified major version.